platform/x86/amd/hsmp: Move structure and macros to header file
commit 1757d2b8dcf6810d7356d8420606df5080e36260 upstream.

This is in preparation to splitting ACPI and platform device drivers.
No logical change, move common structures, macros to hsmp.h,
add missed header files, remove unwanted header inclusions,
and re-order the header file in hsmp.c file.

Link: https://lore.kernel.org/r/20241021111428.2676884-4-suma.hegde@amd.com
